move getReportIcons to reportUtils
